Add ScintillaEx to AppTemplate and move UIActionForm to AppTemplate.
