+++ Sandra.UI/SyntaxEditor.cs
@@ -21,30 +21,19 @@
 
 using Eutherion.Text;
 using Eutherion.UIActions;
-using Eutherion.Utils;
 using Eutherion.Win.AppTemplate;
-using Eutherion.Win.Storage;
-using Eutherion.Win.UIActions;
-using Eutherion.Win.Utils;
 using ScintillaNET;
-using System;
 using System.Drawing;
 using System.Linq;
-using System.Windows.Forms;
 
 namespace Sandra.UI
 {
     /// <summary>
-    /// Represents a <see cref="Scintilla"/> control with syntax highlighting, a number of <see cref="UIAction"/> hooks,
+    /// Represents a <see cref="ScintillaEx"/> control with syntax highlighting, a number of <see cref="UIAction"/> hooks,
     /// and a mouse-wheel event handler.
     /// </summary>
-    public abstract partial class SyntaxEditor<TTerminal> : Scintilla, IUIActionHandlerProvider
+    public abstract class SyntaxEditor<TTerminal> : ScintillaEx
     {
-        /// <summary>
-        /// Gets the action handler for this control.
-        /// </summary>
-        public UIActionHandler ActionHandler { get; } = new UIActionHandler();
-
         protected readonly TextIndex<TTerminal> TextIndex;
 
         protected Style DefaultStyle => Styles[Style.Default];
@@ -63,25 +52,10 @@ namespace Sandra.UI
         {
             TextIndex = new TextIndex<TTerminal>();
 
-            UsePopup(false);
-
-            BufferedDraw = false;
-            Technology = Technology.DirectWrite;
-
             HScrollBar = false;
             VScrollBar = true;
 
             Margins.ForEach(x => x.Width = 0);
-
-            //https://notepad-plus-plus.org/community/topic/12576/list-of-all-assigned-keyboard-shortcuts/8
-            //https://scintilla.org/CommandValues.html
-            //https://sourceforge.net/p/scintilla/code/ci/default/tree/src/KeyMap.h
-            //https://sourceforge.net/p/scintilla/code/ci/default/tree/src/KeyMap.cxx
-            ClearCmdKey(Keys.Control | Keys.F);
-            ClearCmdKey(Keys.Control | Keys.H);
-            ClearCmdKey(Keys.Control | Keys.L);
-            ClearCmdKey(Keys.Control | Keys.R);
-            ClearCmdKey(Keys.Control | Keys.U);
 
             if (Session.Current.TryGetAutoSaveValue(SettingKeys.Zoom, out int zoomFactor))
             {
@@ -98,148 +72,11 @@ namespace Sandra.UI
             }
         }
 
-        /// <summary>
-        /// Occurs when the zoom factor of this <see cref="SyntaxEditor{TTerminal}"/> is updated.
-        /// </summary>
-        public event EventHandler<ZoomFactorChangedEventArgs> ZoomFactorChanged;
-
-        /// <summary>
-        /// Raises the <see cref="ZoomFactorChanged"/> event.
-        /// </summary>
-        /// <param name="e">
-        /// The data for the event.
-        /// </param>
-        protected virtual void OnZoomFactorChanged(ZoomFactorChangedEventArgs e)
-        {
-            ZoomFactorChanged?.Invoke(this, e);
-        }
-
-        private void RaiseZoomFactorChanged(ZoomFactorChangedEventArgs e)
+        protected override void OnZoomFactorChanged(ZoomFactorChangedEventArgs e)
         {
             // Not only raise the event, but also save the zoom factor setting.
-            OnZoomFactorChanged(e);
+            base.OnZoomFactorChanged(e);
             Session.Current.AutoSave.Persist(SettingKeys.Zoom, e.ZoomFactor);
         }
-
-        protected override void OnMouseWheel(MouseEventArgs e)
-        {
-            base.OnMouseWheel(e);
-
-            if (ModifierKeys.HasFlag(Keys.Control))
-            {
-                // Zoom isn't updated yet, so predict here what it's going to be.
-                int newZoomFactorPrediction = Zoom + Math.Sign(e.Delta);
-                if (ScintillaZoomFactor.MinDiscreteValue <= newZoomFactorPrediction
-                    && newZoomFactorPrediction <= ScintillaZoomFactor.MaxDiscreteValue)
-                {
-                    RaiseZoomFactorChanged(new ZoomFactorChangedEventArgs(newZoomFactorPrediction));
-                }
-            }
-        }
-
-        /// <summary>
-        /// Binds the regular cut/copy/paste/select all UIActions to this textbox.
-        /// </summary>
-        public void BindStandardEditUIActions()
-        {
-            var editBindings = new UIActionBindings
-            {
-                { SharedUIAction.Undo, TryUndo },
-                { SharedUIAction.Redo, TryRedo },
-
-                { SharedUIAction.ZoomIn, TryZoomIn },
-                { SharedUIAction.ZoomOut, TryZoomOut },
-
-                { SharedUIAction.CutSelectionToClipBoard, TryCutSelectionToClipBoard },
-                { SharedUIAction.CopySelectionToClipBoard, TryCopySelectionToClipBoard },
-                { SharedUIAction.PasteSelectionFromClipBoard, TryPasteSelectionFromClipBoard },
-                { SharedUIAction.SelectAllText, TrySelectAllText },
-            };
-
-            ShortcutKeysUIActionInterface shortcutKeysInterface = null;
-            editBindings
-                .Where(x => x.Interfaces.TryGet(out shortcutKeysInterface) && shortcutKeysInterface.Shortcuts != null)
-                .SelectMany(x => shortcutKeysInterface.Shortcuts)
-                .Select(KeyUtilities.ToKeys)
-                .ForEach(ClearCmdKey);
-
-            ClearCmdKey(Keys.Oemplus | Keys.Shift | Keys.Control);
-            ClearCmdKey(Keys.OemMinus | Keys.Control);
-
-            this.BindActions(editBindings);
-        }
-    }
-
-    /// <summary>
-    /// Contains extension methods to interface between standard System.Windows.Forms and Scintilla classes.
-    /// </summary>
-    public static class ScintillaExtensions
-    {
-        /// <summary>
-        /// Copies a <see cref="Font"/> definition to a Scintilla <see cref="Style"/>.
-        /// </summary>
-        /// <param name="font">
-        /// The font to copy.
-        /// </param>
-        /// <param name="style">
-        /// The style to copy to.
-        /// </param>
-        /// <exception cref="ArgumentNullException">
-        /// <paramref name="font"/> and/or <paramref name="style"/> are null.
-        /// </exception>
-        public static void CopyTo(this Font font, Style style)
-        {
-            style.Font = font.FontFamily.Name;
-            style.SizeF = font.Size;
-            style.Bold = font.Bold;
-            style.Italic = font.Italic;
-            style.Underline = font.Underline;
-        }
-    }
-
-    /// <summary>
-    /// Provides data for the <see cref="SyntaxEditor{TTerminal}.ZoomFactorChanged"/> event.
-    /// </summary>
-    public class ZoomFactorChangedEventArgs : EventArgs
-    {
-        /// <summary>
-        /// Gets the new zoom factor, represented as an integer in the range [-10..20].
-        /// </summary>
-        public int ZoomFactor { get; }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ZoomFactorChangedEventArgs"/> class.
-        /// </summary>
-        /// <param name="zoomFactor">
-        /// The new zoom factor, represented as an integer in the range [-10..20].
-        /// </param>
-        public ZoomFactorChangedEventArgs(int zoomFactor)
-        {
-            ZoomFactor = zoomFactor;
-        }
-    }
-
-    public sealed class ScintillaZoomFactor : PType.Derived<PInteger, int>
-    {
-        /// <summary>
-        /// Returns the minimum recommended value for the zoom factor
-        /// of a <see cref="Scintilla"/> control which is between -10 and 20, endpoints included.
-        /// </summary>
-        public static readonly int MinDiscreteValue = -10;
-
-        /// <summary>
-        /// Returns the maximum recommended value for the zoom factor
-        /// of a <see cref="Scintilla"/> control which is between -10 and 20, endpoints included.
-        /// </summary>
-        public static readonly int MaxDiscreteValue = 20;
-
-        public static readonly ScintillaZoomFactor Instance = new ScintillaZoomFactor();
-
-        private ScintillaZoomFactor() : base(new PType.RangedInteger(MinDiscreteValue, MaxDiscreteValue)) { }
-
-        public override Union<ITypeErrorBuilder, int> TryGetTargetValue(PInteger integer)
-            => ValidValue((int)integer.Value);
-
-        public override PInteger GetBaseValue(int value) => new PInteger(value);
     }
 }
